BUG: Fix how __main__ handles output_dir_path arg
Because the --output_dir_path arg uses narg='+',
it will always be a list, even if there's one output_dir.
This circumvents logic in `prep_and_save` to catch
the case where a user passes only one output_dir,
that duplicates the string or pathlib.Path to be a list as
long as the `dir_path` list.

We don't want `prep_and_save` to have to know about what
the argparser does, so we catch this one-element list
inside `main` and change it to just the single string,
and then let `prep_and_save` duplicate it
for every `dir_path`.