CLN: Refactor `convert_syl_to_psd` to save compute
Instead of computing PSDS for every syllable,
and then only keeping `max_num_psds`
(now renamed `max_syllables`),
we first make a flattened list of syllables
and then truncate that to `max_syllables`.
This should make things faster in some cases since
computing the PSD is the expensive step,
and we now avoid doing it when unnecessary.